Fix 18 double-to-float compilation errors in Devil's Dream attacks
Changed local `size` variables from double to float in
NightmareConstructs.FloatingRoom and DreamDistortions.FoldingCube.
Added explicit cast in VoidIntrusions.ShadowTentacle.

+++ src/main/java/com/blockforge/chaoscraft/modes/devilsdream/attacks/DreamDistortions.java
@@ -583,7 +583,7 @@ public final class DreamDistortions {
 
             Material[] mats = {Material.QUARTZ_BLOCK, Material.PRISMARINE};
             // 6 faces × 2 blocks each
-            double size = 2.0;
+            float size = 2.0f;
             // Top/Bottom
             faces.add(spawnFace(cubeCenter, 0, size, 0, size * 2, 0.3f, size * 2, mats[0]));
             faces.add(spawnFace(cubeCenter, 0, size, 0, size, 0.3f, size, mats[1]));
